refactor(model): split registry from downloads
