+++ src/extraArches/fastlinedarken.py
@@ -5,7 +5,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 
 from src.infra.isCudaInit import CudaChecker
-from src.model.downloadModels import weightsDir
+from src.model.registry import weightsDir
 
 checker = CudaChecker()
