+++ src/interpolate/rife_ncnn.py
@@ -2,10 +2,8 @@ import numpy as np
 import torch
 
 from src.infra.isCudaInit import CudaChecker
-from src.model.downloadModels import (
-    modelsMap,
-    resolveWeightPath,
-)
+from src.model.download import resolveWeightPath
+from src.model.registry import modelsMap
 
 checker = CudaChecker()
