+++ tests/test_presetLogic.py
@@ -1,4 +1,4 @@
-"""Tests for src/utils/presetLogic.py — --preset save/load round-trip.
+"""Tests for src.server.presetLogic — --preset save/load round-trip.
 
 createPreset is dual-purpose: an unknown preset name SAVES the current args
 (minus the per-run I/O fields), a known name LOADS the stored values onto the
